+++ crates/templated_uri/src/base_uri/base_path.rs
@@ -199,14 +199,14 @@ mod test {
     }
 
     #[test]
-    fn try_from_path_and_query_ref() {
+    fn try_from_path_ref() {
         let paq = PathAndQuery::from_static("/ref/path/");
         let path = BasePath::try_from(&paq).unwrap();
         assert_eq!(path.as_str(), "/ref/path/");
     }
 
     #[test]
-    fn try_from_path_and_query_ref_invalid() {
+    fn try_from_path_ref_invalid() {
         let paq: PathAndQuery = "/no-trailing-slash".parse().unwrap();
         let err = BasePath::try_from(&paq).unwrap_err();
         assert!(err.to_string().contains("the path must start and end with a slash"));